+++ jdaviz/configs/cubeviz/plugins/viewers.py
@@ -109,7 +109,7 @@ class CubevizImageView(JdavizViewerMixin, BqplotImageView):
             #  data for this application. This section will need to be updated
             #  when that is no longer true.
             # Hack to insert WCS for generated 2D and 3D images using FLUX cube WCS.
-            if 'Plugin' in image.meta:
+            if 'Plugin' in image.meta and not image.coords:
                 coo_data = self.jdaviz_app.data_collection[0]
             else:
                 coo_data = image